fixed the ship's movement. it now indeed moves and updates its position, instead of just standing still. it also now can move both left and dright, instead of just right. the problem lay in the fact that 'bank left' was just never called. now I need to unify the movement of both the meteor and ship to both work using their inherited 'Update' method

+++ src/ship.h
@@ -47,7 +47,7 @@ protected:
   void SetShipBase();
   
   // Change Stuff
-  void UpdateInformation(float x_delta);
+  void UpdateInformation(int x_delta);
 
 };
